requirements.txt
carla==0.9.15
antlr4-python3-runtime==4.8
dictor==0.1.7
gql==0.2.0
graphql-core==1.1
hydra-core==1.0.3
imgaug==0.4.0
importlib-resources==3.0.0
matplotlib==3.3.1
omegaconf
opencv-python==4.5.1.48
py-trees==0.8.3
pydot==1.4.1
pygame
pygccxml==2.0.1
pyplusplus==1.8.4
pywavelets==1.1.1
scikit-image==0.17.2
stable-baselines3==0.8.0
subprocess32==3.5.4
tifffile==2020.10.1
typing-extensions==3.7.4.3
wandb
h5py
